+++ grafana-plugin/src/containers/Rotations/ScheduleOverrides.tsx
@@ -10,10 +10,12 @@ import Text from 'components/Text/Text';
 import TimelineMarks from 'components/TimelineMarks/TimelineMarks';
 import Rotation from 'containers/Rotation/Rotation';
 import ScheduleOverrideForm from 'containers/RotationForm/ScheduleOverrideForm';
+import { WithPermissionControl } from 'containers/WithPermissionControl/WithPermissionControl';
 import { getOverrideColor, getOverridesFromStore } from 'models/schedule/schedule.helpers';
 import { Schedule, Shift, ShiftEvents } from 'models/schedule/schedule.types';
 import { Timezone } from 'models/timezone/timezone.types';
 import { WithStoreProps } from 'state/types';
+import { UserAction } from 'state/userAction';
 import { withMobXProviderContext } from 'state/withStore';
 
 import { DEFAULT_TRANSITION_TIMEOUT } from './Rotations.config';
@@ -78,9 +80,11 @@ class ScheduleOverrides extends Component<ScheduleOverridesProps, ScheduleOverri
                   Overrides
                 </Text.Title>
               </div>
-              <Button disabled={disabled} icon="plus" onClick={this.handleAddOverride} variant="secondary">
-                Add override
-              </Button>
+              <WithPermissionControl userAction={UserAction.UpdateSchedules}>
+                <Button disabled={disabled} icon="plus" onClick={this.handleAddOverride} variant="secondary">
+                  Add override
+                </Button>
+              </WithPermissionControl>
             </HorizontalGroup>
           </div>
           <div className={cx('header-plus-content')}>